change output type of wall time to double

+++ src/solver/base_solver.h
@@ -12,7 +12,7 @@
  */
 class SolverResult {
 private:
-  long int t;
+    double t;
   short s;
 
 public:
